Fix segfault in SoundTouch during Change Pitch effect. Previously fixed in our copy of SoundTouch 1.0.  Forwarded upstream.

+++ lib-src/soundtouch/source/SoundTouch/RateTransposer.cpp
@@ -8,10 +8,10 @@
  * Author e-mail : oparviai @ iki.fi
  * File created  : 13-Jan-2002
  * 
- * Last changed  : $Date: 2003/12/27 10:00:52 $
- * File revision : $Revision: 1.7 $
+ * Last changed  : $Date: 2004/03/14 15:51:43 $
+ * File revision : $Revision: 1.1.1.1 $
  *
- * $Id: RateTransposer.cpp,v 1.7 2003/12/27 10:00:52 Olli Exp $
+ * $Id: RateTransposer.cpp,v 1.1.1.1 2004/03/14 15:51:43 mbrubeck Exp $
  *
  * License :
  * 
@@ -257,7 +257,7 @@ void RateTransposer::upsample(const SAMPLETYPE *src, uint numSamples)
 
     // First check that there's enough room in 'storeBuffer' 
     // (+16 is to reserve some slack in the destination buffer)
-    sizeTemp = SCALE * numSamples / uRate + 16;
+    sizeTemp = (int)(SCALE * (double)numSamples / uRate + 16);
 
     // Transpose the samples, store the result into the end of "storeBuffer"
     count = transpose(storeBuffer.ptrEnd(sizeTemp), src, numSamples);
@@ -301,7 +301,7 @@ void RateTransposer::downsample(const SAMPLETYPE *src, uint numSamples)
     storeBuffer.receiveSamples(count);
 
     // Transpose the samples (+16 is to reserve some slack in the destination buffer)
-    sizeTemp = SCALE * numSamples / uRate + 16;
+    sizeTemp = (int)(SCALE * (double)numSamples / uRate + 16);
     count = transpose(outputBuffer.ptrEnd(sizeTemp), tempBuffer.ptrBegin(), count);
     outputBuffer.putSamples(count);
 }
